requirements.txt
tensorflow_datasets~=4.2.0
numpy~=1.18.5
flask~=1.1.2
python-dotenv~=0.15.0
yaspin~=1.4.1
celery~=5.0.5
requests~=2.25.0
kombu~=5.0.2
boto3~=1.17.29
botocore~=1.20.29
tqdm~=4.59.0
progressbar~=2.5
wandb~=0.10.22
python-benedict~=0.24.0
